Refactor(command): remove unused stuff
