+++ src/main/java/net/kodehawa/mantarobot/core/command/compat/Command.java
@@ -15,40 +15,21 @@
  *
  */
 
-package net.kodehawa.mantarobot.core.modules.commands;
+package net.kodehawa.mantarobot.core.command.compat;
 
 import net.kodehawa.mantarobot.core.command.helpers.CommandCategory;
 import net.kodehawa.mantarobot.core.command.helpers.CommandPermission;
 import net.kodehawa.mantarobot.core.command.helpers.HelpContent;
+import net.kodehawa.mantarobot.core.command.helpers.IContext;
 
 import java.util.List;
 
-/**
- * Interface used for handling commands within the bot.
- */
 public interface Command {
-    /**
-     * The Command's {@link CommandCategory}
-     *
-     * @return a Nullable {@link CommandCategory}. Null means that the command should be hidden from Help.
-     */
     CommandCategory category();
-
     CommandPermission permission();
-
-    /**
-     * Invokes the command to be executed.
-     *
-     * @param context     the context of the event that triggered the command
-     * @param commandName the command name that was used
-     * @param content     the arguments of the command
-     */
-    void run(Context context, String commandName, String content);
-
+    void run(IContext context, String commandName, String content);
     HelpContent help();
-
     List<String> getAliases();
-
     default boolean isOwnerCommand() {
         return permission() == CommandPermission.OWNER;
     }
