+++ src/main/java/net/kodehawa/mantarobot/options/core/Option.java
@@ -17,7 +17,7 @@
 
 package net.kodehawa.mantarobot.options.core;
 
-import net.kodehawa.mantarobot.core.command.TextContext;
+import net.kodehawa.mantarobot.core.command.text.TextContext;
 
 import java.util.ArrayList;
 import java.util.HashMap;
