+++ src/main/java/net/kodehawa/mantarobot/utils/commands/FinderUtils.java
@@ -24,7 +24,7 @@ import net.dv8tion.jda.api.entities.channel.concrete.VoiceChannel;
 import net.dv8tion.jda.api.entities.channel.middleman.GuildChannel;
 import net.dv8tion.jda.api.entities.channel.middleman.StandardGuildMessageChannel;
 import net.dv8tion.jda.api.utils.cache.SnowflakeCacheView;
-import net.kodehawa.mantarobot.core.command.slash.IContext;
+import net.kodehawa.mantarobot.core.command.helpers.IContext;
 
 import java.util.ArrayList;
 import java.util.Collections;
